src/client/requirements.txt
Pillow
pygame